requirements.txt
# Peak Load Forecasting Service - Requirements
# Python 3.8+

# Core data processing
pandas>=1.5.0
numpy>=1.23.0

# API clients
requests>=2.28.0
urllib3>=1.26.0

# Machine learning
scikit-learn>=1.2.0

# Data visualization
matplotlib>=3.6.0

# Date/time utilities
python-dateutil>=2.8.0

# Configuration
PyYAML>=6.0

# Optional but recommended
jupyter>=1.0.0  # For notebooks
notebook>=6.5.0  # For running notebooks

